style: 전체적인 컨벤션 수정
