+++ OpenMarket/OpenMarket/Extension/Data+Extension.swift
@@ -1,0 +1,15 @@
+import UIKit.NSDataAsset
+
+extension Data {
+    mutating func append(_ string: String) {
+        if let data = string.data(using: .utf8) {
+            self.append(data)
+        }
+    }
+    
+    mutating func append(_ data: Data?) {
+        if let data = data {
+            self.append(data)
+        }
+    }
+}
