+++ OpenMarket/OpenMarket/Extension/UISegmentControl+Extension.swift
@@ -2,16 +2,12 @@ import UIKit
 
 extension UISegmentedControl {
     func setDefaultFontColor(_ color: UIColor = UIColor.white) {
-        let defaultAttributes = [
-            NSAttributedString.Key.foregroundColor: color
-        ]
+        let defaultAttributes = [NSAttributedString.Key.foregroundColor: color]
         setTitleTextAttributes(defaultAttributes, for: .normal)
     }
 
     func setSelectedFontColor(_ color: UIColor = UIColor.red) {
-        let selectedAttributes = [
-            NSAttributedString.Key.foregroundColor: color
-        ]
+        let selectedAttributes = [NSAttributedString.Key.foregroundColor: color]
         setTitleTextAttributes(selectedAttributes, for: .selected)
     }
 }
