+++ OpenMarket/OpenMarket/Models/DecodableTypes/PostResponse.swift
@@ -1,0 +1,33 @@
+struct PostResponse: Decodable {
+    let id: Int
+    let vendorId: Int
+    let name: String
+    let description: String
+    let thumbnail: String
+    let currency: String
+    let price: Int
+    let bargainPrice: Int
+    let discountedPrice: Int
+    let stock: Int
+    let images: [Image]
+    let vendors: Vendor
+    let createdAt: String
+    let issuedAt: String
+    
+    enum CodingKeys: String, CodingKey, CaseIterable {
+        case id
+        case vendorId = "vendor_id"
+        case name
+        case description
+        case thumbnail
+        case currency
+        case price
+        case bargainPrice = "bargain_price"
+        case discountedPrice = "discounted_price"
+        case stock
+        case images
+        case vendors
+        case createdAt = "created_at"
+        case issuedAt = "issued_at"
+    }
+}
