+++ OpenMarket/OpenMarket/Models/DecodableTypes/Products.swift
@@ -1,0 +1,23 @@
+struct Products: Decodable {
+    let pageNo: Int
+    let itemsPerPage: Int
+    let totalCount: Int
+    let offset: Int
+    let limit: Int
+    let pages: [Page]
+    let lastPage: Int
+    let hasNext: Bool
+    let hasPrev: Bool
+    
+    enum CodingKeys: String, CodingKey, CaseIterable {
+        case pageNo = "page_no"
+        case itemsPerPage = "items_per_page"
+        case totalCount = "total_count"
+        case offset
+        case limit
+        case pages
+        case lastPage = "last_page"
+        case hasNext = "has_next"
+        case hasPrev = "has_prev"
+    }
+}
