+++ OpenMarket/OpenMarket/Models/Enum/DetailViewTitle.swift
