--- OpenMarket/OpenMarket/Models/Products.swift
@@ -1,136 +1,0 @@
-import Foundation
-
-struct Image: Decodable {
-    let id: Int
-    let url: String
-    let thumbnailUrl: String
-    let succeed: Bool
-    let issuedAt: String
-    
-    enum CodingKeys: String, CodingKey, CaseIterable {
-        case id
-        case url
-        case thumbnailUrl = "thumbnail_url"
-        case succeed
-        case issuedAt = "issued_at"
-    }
-}
-
-struct Vendor: Decodable {
-    let name: String
-    let id: Int
-    let createdAt: String
-    let issuedAt: String
-    
-    enum CodingKeys: String, CodingKey, CaseIterable {
-        case name
-        case id
-        case createdAt = "created_at"
-        case issuedAt = "issued_at"
-    }
-}
-
-struct PostResponse: Decodable {
-    let id: Int
-    let vendorId: Int
-    let name: String
-    let description: String
-    let thumbnail: String
-    let currency: String
-    let price: Int
-    let bargainPrice: Int
-    let discountedPrice: Int
-    let stock: Int
-    let images: [Image]
-    let vendors: Vendor
-    let createdAt: String
-    let issuedAt: String
-    
-    enum CodingKeys: String, CodingKey, CaseIterable {
-        case id
-        case vendorId = "vendor_id"
-        case name
-        case description
-        case thumbnail
-        case currency
-        case price
-        case bargainPrice = "bargain_price"
-        case discountedPrice = "discounted_price"
-        case stock
-        case images
-        case vendors
-        case createdAt = "created_at"
-        case issuedAt = "issued_at"
-    }
-}
-
-struct Images: Decodable, Hashable {
-    let id: Int
-    let url: String
-    let thumbnailUrl: String
-    let succeed: Bool
-    let issuedAt: String
-    enum CodingKeys: String, CodingKey, CaseIterable {
-        case id
-        case url
-        case thumbnailUrl = "thumbnail_url"
-        case succeed
-        case issuedAt = "issued_at"
-    }
-}
-
-struct Page: Decodable, Hashable {
-    let id: Int
-    let vendorId: Int
-    let name: String
-    let description: String?
-    let thumbnail: String
-    let currency: String
-    let price: Double
-    let bargainPrice: Double
-    let discountedPrice: Double
-    let stock: Int
-    let images: [Images]?
-    let createdAt: String
-    let issuedAt: String
-    
-    enum CodingKeys: String, CodingKey, CaseIterable {
-        case id
-        case vendorId = "vendor_id"
-        case name
-        case description
-        case thumbnail
-        case currency
-        case price
-        case bargainPrice = "bargain_price"
-        case discountedPrice = "discounted_price"
-        case stock
-        case images
-        case createdAt = "created_at"
-        case issuedAt = "issued_at"
-    }
-}
-
-struct Products: Decodable {
-    let pageNo: Int
-    let itemsPerPage: Int
-    let totalCount: Int
-    let offset: Int
-    let limit: Int
-    let pages: [Page]
-    let lastPage: Int
-    let hasNext: Bool
-    let hasPrev: Bool
-    
-    enum CodingKeys: String, CodingKey, CaseIterable {
-        case pageNo = "page_no"
-        case itemsPerPage = "items_per_page"
-        case totalCount = "total_count"
-        case offset
-        case limit
-        case pages
-        case lastPage = "last_page"
-        case hasNext = "has_next"
-        case hasPrev = "has_prev"
-    }
-}
