+++ OpenMarket/OpenMarket/Models/SendUpdateDelegate.swift
@@ -1,0 +1,3 @@
+protocol SendUpdateDelegate {
+    func sendUpdate() -> UserIdentifier
+}
