+++ OpenMarket/OpenMarket/Views/ItemCollectionViewCell.swift
@@ -276,10 +276,6 @@ extension ItemCollectionViewCell {
     func setImage(by image: UIImage) {
         self.itemImageView.image = image
     }
-    
-    private func setId() -> Int? {
-        return product?.id
-    }
 }
 
 // MARK: - prepare For Reuse
